Tests and improvments of the new LISTEN command (JIRA: CXX-3412)
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@55305 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/netschedule/ns_notifications.cpp
@@ -281,7 +281,7 @@ void CNSNotificationList::NotifyJobStatus(unsigned int    address,
              k_MessageBufferSize - m_JobStateConstPartLength - job_key.size(),
              "&job_status=%s&last_event_index=%ld",
              CNetScheduleAPI::StatusToString(job_status).c_str(),
-             last_event_index);
+             (long int)last_event_index);
 
 
     m_StatusNotificationSocket.Send(
